requirements.txt
beautifulsoup4==4.6.0
python-slugify==1.2.4
requests==2.18.1
simplejson==3.10.0
PyYAML==3.12
prompt==0.4.1
termcolor
